% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{annotated_par_table}
\alias{annotated_par_table}
\title{Build Annotated SAM Parameter Table}
\usage{
annotated_par_table(sam_fit)
}
\arguments{
\item{sam_fit}{A fitted SAM object.}
}
\value{
A tibble of parameter estimates with fleet/age annotations.
}
\description{
Build Annotated SAM Parameter Table
}
