% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{compare_configs}
\alias{compare_configs}
\title{Fit and Score a Candidate SAM Configuration}
\usage{
compare_configs(
  x,
  conf_list = NULL,
  dat = NULL,
  trysim = FALSE,
  allLogN0 = FALSE,
  config_mutator = NULL,
  eval_max = 3000,
  retro_year = 5,
  ncores = 1
)
}
\arguments{
\item{x}{Configuration key/index in \code{conf_list}.}

\item{conf_list}{List of candidate SAM configurations.}

\item{dat}{SAM data object used for fitting.}

\item{trysim}{Whether to run simulation-based convergence checks.}

\item{allLogN0}{Whether to set all \code{keyVarLogN} entries to 0.}

\item{config_mutator}{Optional function \verb{(conf, x)} to mutate config before fit.}

\item{eval_max}{Optimizer maximum evaluations.}

\item{retro_year}{Number of retrospective peels.}

\item{ncores}{Number of cores for parallel steps.}
}
\value{
A tibble with fit diagnostics (AIC, convergence, Mohn metrics).
}
\description{
Fit and Score a Candidate SAM Configuration
}
\examples{
\dontrun{
res <- compare_configs(
  x = "base",
  conf_list = conf_list,
  dat = dat,
  config_mutator = function(conf, id) conf
)
print(res)
}
}
