% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{format_sam_pres}
\alias{format_sam_pres}
\title{Format Process Residuals}
\usage{
format_sam_pres(res)
}
\arguments{
\item{res}{Output from \code{procres(sam_fit)}.}
}
\value{
A tibble with process residual diagnostics.
}
\description{
Format Process Residuals
}
