% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{format_sam_res}
\alias{format_sam_res}
\title{Format Observation Residuals}
\usage{
format_sam_res(res)
}
\arguments{
\item{res}{Output from \code{residuals(sam_fit)}.}
}
\value{
A tibble with observation residual diagnostics.
}
\description{
Format Observation Residuals
}
