% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{input_data_plot}
\alias{input_data_plot}
\title{Plot Input Data by Year and Age}
\usage{
input_data_plot(dat)
}
\arguments{
\item{dat}{Long-format data including \code{year}, \code{age}, and value columns.}
}
\value{
A ggplot object.
}
\description{
Plot Input Data by Year and Age
}
