% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{model_combfit}
\alias{model_combfit}
\title{Plot Combined Survey Observed vs Predicted Fits}
\usage{
model_combfit(
  res,
  model_dat,
  min_year = NULL,
  excluded_fleets = c("Residual catch"),
  reference_fleet = "spring",
  age_shift_other_fleets = 1,
  year_shift_other_fleets = 1,
  max_age = 12,
  colors = NULL
)
}
\arguments{
\item{res}{Result list with \verb{$res}.}

\item{model_dat}{Model input data.}

\item{min_year}{Minimum year to include.}

\item{excluded_fleets}{Fleets excluded from plotting.}

\item{reference_fleet}{Fleet used for no age/year shifting.}

\item{age_shift_other_fleets}{Age shift for non-reference fleets.}

\item{year_shift_other_fleets}{Year shift for non-reference fleets.}

\item{max_age}{Maximum age after shifting.}

\item{colors}{Optional color vector for fleets.}
}
\value{
A ggplot object.
}
\description{
Plot Combined Survey Observed vs Predicted Fits
}
