% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{model_data_plot}
\alias{model_data_plot}
\title{Plot Scaled Input Data Sources Over Time}
\usage{
model_data_plot(res = NULL, model_dat = NULL, maturation_cutoff_year = NULL)
}
\arguments{
\item{res}{Unused legacy argument.}

\item{model_dat}{Model input data.}

\item{maturation_cutoff_year}{Year before which maturity/weights are set to zero.}
}
\value{
A ggplot object.
}
\description{
Plot Scaled Input Data Sources Over Time
}
