% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{model_ices_plot}
\alias{model_ices_plot}
\title{Plot Main SAM Time-Series with Optional Observed Catch Overlay}
\usage{
model_ices_plot(
  res,
  model_dat,
  observed_catch = TRUE,
  observed_catch_start_year = NULL,
  observed_catch_end_year = NULL
)
}
\arguments{
\item{res}{Result list with \verb{$fit}.}

\item{model_dat}{Model input data.}

\item{observed_catch}{Whether to add observed catch points.}

\item{observed_catch_start_year}{First year for observed catch overlay.}

\item{observed_catch_end_year}{Last year for observed catch overlay.}
}
\value{
A ggplot object.
}
\description{
Plot Main SAM Time-Series with Optional Observed Catch Overlay
}
