% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{model_lik_plot}
\alias{model_lik_plot}
\title{Plot Likelihood-Related Parameter Diagnostics}
\usage{
model_lik_plot(res, cor_fit = NULL)
}
\arguments{
\item{res}{Result list with \verb{$fit} and optional \verb{$joined_F$fit}.}

\item{cor_fit}{Optional model fit to use for correlation parameters.}
}
\value{
A patchwork/ggplot composite.
}
\description{
Plot Likelihood-Related Parameter Diagnostics
}
