% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{model_lo_plot}
\alias{model_lo_plot}
\title{Plot Leave-One-Out Comparisons}
\usage{
model_lo_plot(res)
}
\arguments{
\item{res}{Result list with \verb{$fit} and \verb{$lo}.}
}
\value{
A ggplot object.
}
\description{
Plot Leave-One-Out Comparisons
}
