% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{model_pres_resid_plot}
\alias{model_pres_resid_plot}
\title{Plot Process Residuals by Fleet and Age}
\usage{
model_pres_resid_plot(res)
}
\arguments{
\item{res}{Result list with \verb{$process_res}.}
}
\value{
A ggplot object.
}
\description{
Plot Process Residuals by Fleet and Age
}
