% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{model_resid_plot}
\alias{model_resid_plot}
\title{Plot Observation Residuals by Fleet and Age}
\usage{
model_resid_plot(res)
}
\arguments{
\item{res}{Result list with \verb{$res}.}
}
\value{
A ggplot object.
}
\description{
Plot Observation Residuals by Fleet and Age
}
