% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{model_retro_plot}
\alias{model_retro_plot}
\title{Plot Retrospective Runs with Mohn's Rho}
\usage{
model_retro_plot(res)
}
\arguments{
\item{res}{Result list with \verb{$fit} and \verb{$retro}.}
}
\value{
A ggplot object.
}
\description{
Plot Retrospective Runs with Mohn's Rho
}
