% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{model_sifit_plot}
\alias{model_sifit_plot}
\title{Plot Survey Index Fits by Fleet and Age}
\usage{
model_sifit_plot(res)
}
\arguments{
\item{res}{Result list with \verb{$res}.}
}
\value{
A ggplot object.
}
\description{
Plot Survey Index Fits by Fleet and Age
}
