% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{rby.sam}
\alias{rby.sam}
\title{Extract Results by Year}
\usage{
rby.sam(sam_fit, run_ref_bio = TRUE, ref_bio_type = "length")
}
\arguments{
\item{sam_fit}{A fitted SAM object.}

\item{run_ref_bio}{Whether to include \code{ref_bio} and \code{hr} derived series.}

\item{ref_bio_type}{Reference biomass type (\code{"length"} or \code{"age"}).}
}
\value{
A tibble with median/lower/upper time series by variable.
}
\description{
Extract Results by Year
}
