% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{rbya.sam}
\alias{rbya.sam}
\title{Extract Results by Year and Age}
\usage{
rbya.sam(
  sam_fit,
  res = NULL,
  include_stock_data = TRUE,
  include_parameter_table = TRUE
)
}
\arguments{
\item{sam_fit}{A fitted SAM object.}

\item{res}{Deprecated/unused legacy argument.}

\item{include_stock_data}{Whether to join stock/catch/maturity/F/M data.}

\item{include_parameter_table}{Whether to join annotated parameter estimates.}
}
\value{
A tibble with year-age level outputs.
}
\description{
Extract Results by Year and Age
}
