% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{ref_bio}
\alias{ref_bio}
\title{Compute Reference Biomass from Year-Age Outputs}
\usage{
ref_bio(fit)
}
\arguments{
\item{fit}{A fitted SAM object.}
}
\value{
A tibble with \code{year} and \code{ref_bio}.
}
\description{
Compute Reference Biomass from Year-Age Outputs
}
