% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{run_sam_dashboard}
\alias{run_sam_dashboard}
\title{Run SAM Dashboard}
\usage{
run_sam_dashboard(
  model_dir = "testmore/nsher",
  retro_year = 5,
  output_file = "sam-dashboard.html",
  output_dir = ".",
  launch.browser = interactive()
)
}
\arguments{
\item{model_dir}{Directory containing NSHER inputs.}

\item{retro_year}{Number of retrospective peels.}

\item{output_file}{Output HTML filename.}

\item{output_dir}{Output directory for rendered HTML.}

\item{launch.browser}{If \code{TRUE}, attempts to open the rendered file.}
}
\value{
Path to rendered HTML dashboard.
}
\description{
Run SAM Dashboard
}
