% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{sam.dir.write}
\alias{sam.dir.write}
\title{Write SAM Input Data to Disk}
\usage{
sam.dir.write(d, md)
}
\arguments{
\item{d}{Data frame to write.}

\item{md}{Model directory path.}
}
\value{
Invisibly returns the written data frame.
}
\description{
Write SAM Input Data to Disk
}
