% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{sam_dashboard_app}
\alias{sam_dashboard_app}
\title{Build SAM Dashboard App}
\usage{
sam_dashboard_app(model_dir = "testmore/nsher", retro_year = 5)
}
\arguments{
\item{model_dir}{Directory containing NSHER inputs.}

\item{retro_year}{Number of retrospective peels.}
}
\value{
A Shiny app object.
}
\description{
Build SAM Dashboard App
}
