% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{sam_dashboard_as_table}
\alias{sam_dashboard_as_table}
\title{Convert SAM Input Object to Table-Friendly Data Frame}
\usage{
sam_dashboard_as_table(x)
}
\arguments{
\item{x}{Object read by \code{stockassessment::read.ices()}.}
}
\value{
A data frame suitable for display in dashboard tables.
}
\description{
Convert SAM Input Object to Table-Friendly Data Frame
}
\keyword{internal}
