% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{sam_dashboard_read_inputs}
\alias{sam_dashboard_read_inputs}
\title{Read NSHER SAM Input Files}
\usage{
sam_dashboard_read_inputs(model_dir)
}
\arguments{
\item{model_dir}{Directory containing NSHER \code{.dat} input files.}
}
\value{
A named list of raw ICES inputs.
}
\description{
Read NSHER SAM Input Files
}
\keyword{internal}
