% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{sam_default_fleet_label}
\alias{sam_default_fleet_label}
\title{Map Fleet IDs to Display Labels}
\usage{
sam_default_fleet_label(fleet, labels = NULL)
}
\arguments{
\item{fleet}{Character vector of fleet ids.}

\item{labels}{Optional named vector/list used for recoding.}
}
\value{
A character vector of fleet labels.
}
\description{
Map Fleet IDs to Display Labels
}
