% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{sam_simulate}
\alias{sam_simulate}
\title{Simulate and Refit SAM}
\usage{
sam_simulate(sam_fit)
}
\arguments{
\item{sam_fit}{A fitted SAM object.}
}
\value{
A \code{samset} list of simulated fits with the original fit as attribute.
}
\description{
Simulate and Refit SAM
}
