% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMutils-package.R
\docType{package}
\name{SAMutils}
\alias{SAMutils-package}
\alias{SAMutils}
\title{SAMutils: Utilities for SAM Model Workflows}
\description{
Utility functions for building SAM input objects, fitting
configurations, and creating diagnostics and plots.
}
\author{
\strong{Maintainer}: SAMutils Authors \email{samutils@example.com}

}
