% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{SR_plot}
\alias{SR_plot}
\title{Plot Stock-Recruit Relationship with Uncertainty}
\usage{
SR_plot(res, ssb_year_offset = 1)
}
\arguments{
\item{res}{Result list with \verb{$fit}.}

\item{ssb_year_offset}{Offset applied to SSB year to align with recruitment.}
}
\value{
A ggplot object.
}
\description{
Plot Stock-Recruit Relationship with Uncertainty
}
