% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{run_sam_dashboard}
\alias{run_sam_dashboard}
\title{Run SAM Dashboard}
\usage{
run_sam_dashboard(
  model_dir = "testmore/nsher",
  retro_year = 5,
  launch.browser = interactive()
)
}
\arguments{
\item{model_dir}{Directory containing NSHER inputs.}

\item{retro_year}{Number of retrospective peels.}

\item{launch.browser}{Passed to \code{shiny::runApp()}.}
}
\value{
Invisibly returns the Shiny app run result.
}
\description{
Run SAM Dashboard
}
